pkg/synnergy_network/core/blockchain/predictive_chain_management/predictive_chain_management.go
package predictivechainmanagement

import (
	"fmt"
	"log"
	"sync"
	"time"

	"github.com/synnergy_network/core/blockchain/consensus"
	"github.com/synnergy_network/core/blockchain/security"
	"github.com/synnergy_network/core/blockchain/utils"
	"github.com/synnergy_network/core/blockchain/validation"
	"github.com/synnergy_network/core/crypto"
)

// PredictiveChainManagement manages predictive modeling and proactive measures
type PredictiveChainManagement struct {
	mu                    sync.Mutex
	networkMetrics        *NetworkMetrics
	predictiveModels      *PredictiveModels
	adaptiveRiskAssess    *AdaptiveRiskAssessment
	securityProtocol      *security.Protocol
	consensusManager      *consensus.Manager
	validationManager     *validation.Manager
	miningPools           *MiningPools
}

// NetworkMetrics holds the real-time metrics of the network
type NetworkMetrics struct {
	BlockCreationTime      time.Duration
	TransactionThroughput  int
	NetworkLatency         time.Duration
	NodePerformance        map[string]float64
}

// PredictiveModels represents the machine learning models used for prediction
type PredictiveModels struct {
	ForkPredictionModel    *utils.MLModel
	ReorganizationModel    *utils.MLModel
	ProfitabilityModel     *utils.MLModel
}

// AdaptiveRiskAssessment dynamically assesses and manages the risk levels
type AdaptiveRiskAssessment struct {
	RiskFactors            map[string]float64
}

// MiningPools manages the decentralized mining pools
type MiningPools struct {
	Pools                  map[string]*MiningPool
}

// MiningPool represents a mining pool
type MiningPool struct {
	ID                     string
	Miners                 map[string]*Miner
}

// Miner represents a miner in the pool
type Miner struct {
	ID                     string
	Performance            float64
}

// NewPredictiveChainManagement initializes the PredictiveChainManagement with the necessary components
func NewPredictiveChainManagement() *PredictiveChainManagement {
	return &PredictiveChainManagement{
		networkMetrics: &NetworkMetrics{
			BlockCreationTime:     time.Duration(0),
			TransactionThroughput: 0,
			NetworkLatency:        time.Duration(0),
			NodePerformance:       make(map[string]float64),
		},
		predictiveModels: &PredictiveModels{
			ForkPredictionModel: utils.NewMLModel("fork"),
			ReorganizationModel: utils.NewMLModel("reorg"),
			ProfitabilityModel:  utils.NewMLModel("profitability"),
		},
		adaptiveRiskAssess: &AdaptiveRiskAssessment{
			RiskFactors: make(map[string]float64),
		},
		securityProtocol: security.NewProtocol(),
		consensusManager: consensus.NewManager(),
		validationManager: validation.NewManager(),
		miningPools: &MiningPools{
			Pools: make(map[string]*MiningPool),
		},
	}
}

// CollectMetrics collects real-time metrics from the network
func (pcm *PredictiveChainManagement) CollectMetrics() {
	pcm.mu.Lock()
	defer pcm.mu.Unlock()

	// Simulate collecting metrics
	pcm.networkMetrics.BlockCreationTime = time.Duration(rand.Intn(100)) * time.Millisecond
	pcm.networkMetrics.TransactionThroughput = rand.Intn(1000)
	pcm.networkMetrics.NetworkLatency = time.Duration(rand.Intn(100)) * time.Millisecond
	for i := 0; i < 10; i++ {
		nodeID := fmt.Sprintf("Node-%d", i)
		pcm.networkMetrics.NodePerformance[nodeID] = rand.Float64()
	}

	log.Println("Metrics collected:", pcm.networkMetrics)
}

// PredictNetworkConditions uses machine learning models to predict network conditions
func (pcm *PredictiveChainManagement) PredictNetworkConditions() {
	pcm.mu.Lock()
	defer pcm.mu.Unlock()

	// Simulate predictions
	forkRisk := pcm.predictiveModels.ForkPredictionModel.Predict(pcm.networkMetrics)
	reorgRisk := pcm.predictiveModels.ReorganizationModel.Predict(pcm.networkMetrics)
	profitability := pcm.predictiveModels.ProfitabilityModel.Predict(pcm.networkMetrics)

	log.Println("Predicted Fork Risk:", forkRisk)
	log.Println("Predicted Reorg Risk:", reorgRisk)
	log.Println("Predicted Profitability:", profitability)
}

// AdjustMiningParameters adjusts the mining parameters based on predictions
func (pcm *PredictiveChainManagement) AdjustMiningParameters() {
	pcm.mu.Lock()
	defer pcm.mu.Unlock()

	// Simulate adjustments
	pcm.consensusManager.AdjustDifficulty(rand.Intn(100))
	pcm.validationManager.AdjustBlockSize(rand.Intn(100))

	log.Println("Mining parameters adjusted")
}

// AssessRisk dynamically assesses and manages the risk levels
func (pcm *PredictiveChainManagement) AssessRisk() {
	pcm.mu.Lock()
	defer pcm.mu.Unlock()

	// Simulate risk assessment
	for k := range pcm.adaptiveRiskAssess.RiskFactors {
		pcm.adaptiveRiskAssess.RiskFactors[k] = rand.Float64()
	}

	log.Println("Risk factors assessed:", pcm.adaptiveRiskAssess.RiskFactors)
}

// ApplySecurityMeasures applies enhanced security measures based on risk assessment
func (pcm *PredictiveChainManagement) ApplySecurityMeasures() {
	pcm.mu.Lock()
	defer pcm.mu.Unlock()

	// Simulate applying security measures
	for _, nodeID := range pcm.securityProtocol.GetNodeIDs() {
		if pcm.adaptiveRiskAssess.RiskFactors[nodeID] > 0.5 {
			pcm.securityProtocol.IsolateNode(nodeID)
		}
	}

	log.Println("Security measures applied")
}

// OptimizeMiningEfforts optimizes mining efforts based on real-time data and predictions
func (pcm *PredictiveChainManagement) OptimizeMiningEfforts() {
	pcm.CollectMetrics()
	pcm.PredictNetworkConditions()
	pcm.AdjustMiningParameters()
	pcm.AssessRisk()
	pcm.ApplySecurityMeasures()
	pcm.DistributeMiningTasks()
	pcm.ContinuousImprovement()
	pcm.AdaptiveLearning()
	pcm.DecentralizedModelTraining()
	pcm.QuantumResistantPrediction()
}

// Start initializes the predictive chain management mechanisms and starts the process
func (pcm *PredictiveChainManagement) Start() {
	go func() {
		for {
			pcm.OptimizeMiningEfforts()
			time.Sleep(1 * time.Minute)
		}
	}()
}

// ContinuousImprovement continuously improves the predictive models based on new data
func (pcm *PredictiveChainManagement) ContinuousImprovement() {
	pcm.mu.Lock()
	defer pcm.mu.Unlock()

	// Simulate continuous improvement
	newData := pcm.networkMetrics
	pcm.predictiveModels.ForkPredictionModel.Update(newData)
	pcm.predictiveModels.ReorganizationModel.Update(newData)
	pcm.predictiveModels.ProfitabilityModel.Update(newData)

	log.Println("Predictive models updated with new data")
}

// AdaptiveLearning dynamically adapts the models to changing network conditions
func (pcm *PredictiveChainManagement) AdaptiveLearning() {
	pcm.mu.Lock()
	defer pcm.mu.Unlock()

	// Simulate adaptive learning
	pcm.predictiveModels.ForkPredictionModel.Adapt(pcm.networkMetrics)
	pcm.predictiveModels.ReorganizationModel.Adapt(pcm.networkMetrics)
	pcm.predictiveModels.ProfitabilityModel.Adapt(pcm.networkMetrics)

	log.Println("Predictive models adapted to changing network conditions")
}

// DecentralizedModelTraining trains predictive models using decentralized data
func (pcm *PredictiveChainManagement) DecentralizedModelTraining() {
	pcm.mu.Lock()
	defer pcm.mu.Unlock()

	// Simulate decentralized model training
	for nodeID, performance := range pcm.networkMetrics.NodePerformance {
		data := pcm.networkMetrics
		data.NodePerformance = map[string]float64{nodeID: performance}
		pcm.predictiveModels.ForkPredictionModel.DecentralizedTrain(data)
		pcm.predictiveModels.ReorganizationModel.DecentralizedTrain(data)
		pcm.predictiveModels.ProfitabilityModel.DecentralizedTrain(data)
	}

	log.Println("Decentralized training of predictive models completed")
}

// QuantumResistantPrediction enhances models with quantum-resistant techniques
func (pcm *PredictiveChainManagement) QuantumResistantPrediction() {
	pcm.mu.Lock()
	defer pcm.mu.Unlock()

	// Simulate integration of quantum-resistant techniques
	pcm.predictiveModels.ForkPredictionModel.ApplyQuantumResistance()
	pcm.predictiveModels.ReorganizationModel.ApplyQuantumResistance()
	pcm.predictiveModels.ProfitabilityModel.ApplyQuantumResistance()

	log.Println("Predictive models enhanced with quantum-resistant techniques")
}

// DistributeMiningTasks distributes mining tasks based on predictive insights
func (pcm *PredictiveChainManagement) DistributeMiningTasks() {
	pcm.mu.Lock()
	defer pcm.mu.Unlock()

	// Simulate task distribution
	for poolID, pool := range pcm.miningPools.Pools {
		for minerID, miner := range pool.Miners {
			profitability := pcm.predictiveModels.ProfitabilityModel.PredictForMiner(pcm.networkMetrics, minerID)
			if profitability > 0.5 {
				// Assign high priority tasks
				log.Printf("Assigning high priority tasks to miner %s in pool %s", minerID, poolID)
			} else {
				// Assign low priority tasks
				log.Printf("Assigning low priority tasks to miner %s in pool %s", minerID, poolID)
			}
		}
	}

	log.Println("Mining tasks distributed")
}

// CreateMiningPool creates a new mining pool
func (pcm *PredictiveChainManagement) CreateMiningPool(poolID string) {
	pcm.mu.Lock()
	defer pcm.mu.Unlock()

	pcm.miningPools.Pools[poolID] = &MiningPool{
		ID:     poolID,
		Miners: make(map[string]*Miner),
	}

	log.Printf("Mining pool %s created", poolID)
}

// AddMinerToPool adds a miner to a specified pool
func (pcm *PredictiveChainManagement) AddMinerToPool(poolID string, minerID string) {
	pcm.mu.Lock()
	defer pcm.mu.Unlock()

	if pool, exists := pcm.miningPools.Pools[poolID]; exists {
		pool.Miners[minerID] = &Miner{
			ID:          minerID,
			Performance: rand.Float64(),
		}
		log.Printf("Miner %s added to pool %s", minerID, poolID)
	} else {
		log.Printf("Mining pool %s does not exist", poolID)
	}
}

// RemoveMinerFromPool removes a miner from a specified pool
func (pcm *PredictiveChainManagement) RemoveMinerFromPool(poolID string, minerID string) {
	pcm.mu.Lock()
	defer pcm.mu.Unlock()

	if pool, exists := pcm.miningPools.Pools[poolID]; exists {
		delete(pool.Miners, minerID)
		log.Printf("Miner %s removed from pool %s", minerID, poolID)
	} else {
		log.Printf("Mining pool %s does not exist", poolID)
	}
}

// IncentivizeMiners provides incentives to miners based on their performance
func (pcm *PredictiveChainManagement) IncentivizeMiners() {
	pcm.mu.Lock()
	defer pcm.mu.Unlock()

	// Simulate incentives
	for poolID, pool := range pcm.miningPools.Pools {
		for minerID, miner := range pool.Miners {
			if miner.Performance > 0.75 {
				// Provide high incentives
				log.Printf("Providing high incentives to miner %s in pool %s", minerID, poolID)
			} else {
				// Provide low incentives
				log.Printf("Providing low incentives to miner %s in pool %s", minerID, poolID)
			}
		}
	}

	log.Println("Incentives provided to miners")
}

